requirements.txt
numpy==2.4.2
sounddevice==0.5.5
faster-whisper==1.2.1
openwakeword==0.6.0
requests==2.32.5
piper-tts==1.4.1
psutil>=6.0.0
pygame>=2.1.0
scipy>=1.11.0
silero-vad>=5.1
